+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## [2.0.0] - 2025-21-10
+## [2.0.0] - 2025-23-10
 
 ### Added
-- Updated sounds for installing/deleting the fusion core and entering/exiting power armor
-- Added client-server synchronization for playing the above sounds
+- The power armor pilot's repulsion system has been improved: the repulsion force depends on the impact force (damage)
+- Other mobs can no longer push the power armor pilot with their collision
+- Fixed an issue where SRP mobs would attack the exoskeleton (entity)